requirements.txt
discord.py[voice]>=2.3.2
PyNaCl>=1.5.0
yt-dlp>=2024.3.10
spotipy>=2.23.0
python-dotenv>=1.0.0
google-generativeai>=0.7.2
requests>=2.31.0
